+++ scripts/ci/update_ext_cmd_tree.py
@@ -46,7 +46,7 @@ def update_cmd_tree(ext_name):
     sys.path.append(ext_dir)
     extension_command_table, _ = _load_extension_command_loader(invoker.commands_loader, None, ext_mod)
 
-    EXT_CMD_TREE_TO_UPLOAD = Session()
+    EXT_CMD_TREE_TO_UPLOAD = Session(encoding='utf-8')
     EXT_CMD_TREE_TO_UPLOAD.load(os.path.expanduser(os.path.join('~', '.azure', file_name)))
     root = {}
     for cmd_name, ext_cmd in extension_command_table.items():
@@ -81,14 +81,21 @@ def upload_cmd_tree():
     file_path = os.path.expanduser(os.path.join('~', '.azure', file_name))
 
     cmd = ['az', 'storage', 'blob', 'upload', '--container-name', f'{STORAGE_CONTAINER}', '--account-name',
-           f'{STORAGE_ACCOUNT}', '--name', f'{blob_file_name}', '--file', f'{file_path}', '--auth-mode', 'login']
-    result = subprocess.run(cmd, check=True, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    print(json.loads(result.stdout))
+           f'{STORAGE_ACCOUNT}', '--name', f'{blob_file_name}', '--file', f'{file_path}', '--auth-mode', 'login',
+           '--overwrite']
+    result = subprocess.run(cmd, capture_output=True)
+    if result.returncode != 0:
+        print(f"Failed to upload '{blob_file_name}' to the storage account")
+        print(result)
 
     cmd = ['az', 'storage', 'blob', 'url', '--container-name', f'{STORAGE_CONTAINER}', '--account-name',
            f'{STORAGE_ACCOUNT}', '--name', f'{blob_file_name}', '--auth-mode', 'login']
-    result = subprocess.run(cmd, check=True, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    url = json.loads(result.stdout)
+    result = subprocess.run(cmd, capture_output=True)
+    if result.stdout and result.returncode == 0:
+        url = json.loads(result.stdout)
+    else:
+        print(f"Failed to get the URL for '{blob_file_name}'")
+        raise
 
     download_file_path = os.path.expanduser(os.path.join('~', '.azure', downloaded_file_name))
     download_file(url, download_file_path)