+++ src/aks-preview/HISTORY.rst
@@ -14,6 +14,7 @@ Pending
 * Vendor new SDK and bump API version to 2023-11-02-preview.
 * Implicitly enable istio when ingress or egress gateway is enabled for Azure Service Mesh.
 * Add parameter to set revision `--revision` for the Azure Service Mesh addon while creating AKS cluster.
+* Add `az aks nodepool delete-machines` command.
 
 1.0.0b5
 +++++++